Added Screening History
Next thing to focus would be allowing user to redo the test when they click the button that takes them to the questionnaire with the exact same setting instead of the form page.

+++ database/migrations/2024_11_11_071151_add_google_id_to_users_table.php
@@ -6,23 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('google_id')->nullable();
+            if (!Schema::hasColumn('users', 'google_id')) {
+                $table->string('google_id')->nullable();
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('google_id');
+            if (Schema::hasColumn('users', 'google_id')) {
+                $table->dropColumn('google_id');
+            }
         });
     }
 };
